docs: 문서 업데이트 및 세팅 수정
