+++ apps/docs/docs/core/index.md
@@ -22,4 +22,4 @@ custom_edit_url: null
 [Issues](https://github.com/mbti-nf-team/frontend-libraries/issues)
 
 ## 🔒 LICENSE
-This libraries is [MIT licensed](https://github.com/mbti-nf-team/frontend-libraries/blob/main/LICENSE).
+This libraries is [MIT licensed](https://github.com/mbti-nf-team/frontend-libraries/blob/main/packages/core/LICENSE).
