+++ apps/docs/docs/react/index.md
@@ -25,4 +25,4 @@ custom_edit_url: null
 [Issues](https://github.com/mbti-nf-team/frontend-libraries/issues)
 
 ## 🔒 LICENSE
-This libraries is [MIT licensed](https://github.com/mbti-nf-team/frontend-libraries/blob/main/LICENSE).
+This libraries is [MIT licensed](https://github.com/mbti-nf-team/frontend-libraries/blob/main/packages/react/LICENSE).
