+++ apps/docs/docusaurus.config.js
@@ -7,7 +7,7 @@ const darkCodeTheme = require('prism-react-renderer/themes/dracula');
 /** @type {import('@docusaurus/types').Config} */
 const config = {
   title: '@nf-team/frontend-libraries',
-  tagline: 'Frontend Libraries',
+  tagline: '@mbti-nf-team 프론트엔드 라이브러리 모음',
   favicon: 'img/favicon.ico',
   url: 'https://mbti-nf-team.github.io',
   baseUrl: '/frontend-libraries/',
@@ -63,14 +63,7 @@ const config = {
           // Please change this to your repo.
           // Remove this to remove the "edit this page" links.
           editUrl:
-            'https://github.com/facebook/docusaurus/tree/main/packages/create-docusaurus/templates/shared/',
-        },
-        blog: {
-          showReadingTime: true,
-          // Please change this to your repo.
-          // Remove this to remove the "edit this page" links.
-          editUrl:
-            'https://github.com/facebook/docusaurus/tree/main/packages/create-docusaurus/templates/shared/',
+            'https://github.com/mbti-nf-team/frontend-libraries/tree/main/apps/docs/',
         },
         theme: {
           customCss: require.resolve('./src/css/custom.css'),
@@ -88,7 +81,7 @@ const config = {
         title: '@nf-team/frontend-libraries',
         logo: {
           alt: 'My Site Logo',
-          src: 'img/logo.svg',
+          src: 'img/logo.png',
         },
         items: [
           {
