+++ apps/docs/src/components/HomepageFeatures/index.tsx
@@ -2,7 +2,7 @@
 /* eslint-disable global-require */
 import React from 'react';
 
-import clsx from 'clsx';
+import Feature from '../Feature';
 
 import styles from './styles.module.css';
 
@@ -10,78 +10,39 @@ type FeatureItem = {
   id: number;
   title: string;
   Svg: React.ComponentType<React.ComponentProps<'svg'>>;
-  description: JSX.Element;
 };
 
 const FeatureList: FeatureItem[] = [
   {
     id: 1,
-    title: 'Easy to Use',
+    title: '@nf-team/core',
     Svg: require('@site/static/img/undraw_docusaurus_mountain.svg').default,
-    description: (
-      <>
-        Docusaurus was designed from the ground up to be easily installed and
-        used to get your website up and running quickly.
-      </>
-    ),
+
   },
   {
     id: 2,
-    title: 'Focus on What Matters',
+    title: '@nf-team/react',
     Svg: require('@site/static/img/undraw_docusaurus_tree.svg').default,
-    description: (
-      <>
-        Docusaurus lets you focus on your docs, and we&apos;ll do the chores. Go
-        ahead and move your docs into the
-        {' '}
-        <code>docs</code>
-        {' '}
-        directory.
-      </>
-    ),
+
   },
   {
     id: 3,
-    title: 'Powered by React',
+    title: '@nf-team/eslint-config',
     Svg: require('@site/static/img/undraw_docusaurus_react.svg').default,
-    description: (
-      <>
-        Extend or customize your website layout by reusing React. Docusaurus can
-        be extended while reusing the same header and footer.
-      </>
-    ),
   },
 ];
 
-function Feature({
-  title, Svg, description, id,
-}: FeatureItem) {
-  return (
-    <div className={clsx('col col--4')}>
-      <div className="text--center">
-        <Svg className={styles.featureSvg} role="img" />
-      </div>
-      <div className="text--center padding-horiz--md">
-        <h3>{title}</h3>
-        <p>{description}</p>
-      </div>
-    </div>
-  );
-}
-
-export default function HomepageFeatures(): JSX.Element {
+function HomepageFeatures(): JSX.Element {
   return (
     <section className={styles.features}>
       <div className="container">
         <div className="row">
           {FeatureList.map(({
-            Svg, description, id, title,
+            Svg, id, title,
           }) => (
             <Feature
               key={id}
               Svg={Svg}
-              description={description}
-              id={id}
               title={title}
             />
           ))}
@@ -90,3 +51,5 @@ export default function HomepageFeatures(): JSX.Element {
     </section>
   );
 }
+
+export default HomepageFeatures;
