+++ apps/docs/src/pages/index.tsx
@@ -1,37 +1,16 @@
-import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import HomepageFeatures from '@site/src/components/HomepageFeatures';
 import Layout from '@theme/Layout';
-import clsx from 'clsx';
 
-import styles from './index.module.css';
+import HomepageHeader from '../components/HomepageHeader';
 
-function HomepageHeader() {
+function Home() {
   const { siteConfig } = useDocusaurusContext();
-  return (
-    <header className={clsx('hero hero--primary', styles.heroBanner)}>
-      <div className="container">
-        <h1 className="hero__title">{siteConfig.title}</h1>
-        <p className="hero__subtitle">{siteConfig.tagline}</p>
-        <div className={styles.buttons}>
-          <Link
-            className="button button--secondary button--lg"
-            to="/docs"
-          >
-            Docusaurus Tutorial - 5min ⏱️
-          </Link>
-        </div>
-      </div>
-    </header>
-  );
-}
 
-export default function Home(): JSX.Element {
-  const { siteConfig } = useDocusaurusContext();
   return (
     <Layout
-      title={`Hello from ${siteConfig.title}`}
-      description="Description will go into a meta tag in <head />"
+      title={siteConfig.title}
+      description="@mbti-nf-team 프론트엔드 라이브러리 모음"
     >
       <HomepageHeader />
       <main>
@@ -40,3 +19,5 @@ export default function Home(): JSX.Element {
     </Layout>
   );
 }
+
+export default Home;
